convert to f64

+++ audio.odin
@@ -21,12 +21,12 @@ Logger :: struct {
 Wav_Data :: struct {
     waveform: wav.Waveform,
     warp_waveform: wav.Waveform,
-    warp_amt: f32,
+    warp_amt: f64,
 }
 
 Note :: struct {
-    frequency: f32,
-    phase: f32,
+    frequency: f64,
+    phase: f64,
     time: f64,
     instrument: int,
 }
@@ -34,7 +34,7 @@ Note :: struct {
 // ----------------------------------------------------------------------------------- helpers
 update_phase :: proc(note: ^Note, sampleRate: u32) {
     
-    note.phase += note.frequency / f32(sampleRate)
+    note.phase += note.frequency / f64(sampleRate)
     if note.phase >= 1 do note.phase -= 1
 
     dt := 1.0 / f64(sampleRate)
@@ -50,19 +50,19 @@ audio_callback :: proc "c" (pDevice: ^ma.device, pOutput, pInput: rawptr, frameC
     audio_data.logger.bufferSize = frameCount
 
     note := &audio_data.note
-    gain : f32 = .1
+    gain : f64 = .1
 
     // ----------------------------------------------------------------------------------- fill output buffer
     for i in 0..<frameCount
     {
-        value : f32
+        value : f64
         if (audio_data.logger.isPlaying)
         {
             value = wav.get_wave_value(note.phase, audio_data.waveData.waveform, audio_data.waveData.warp_waveform, audio_data.waveData.warp_amt)
         }
         
-        output[i * 2]     = value * gain 
-        output[i * 2 + 1] = value * gain
+        output[i * 2]     = f32(value * gain)
+        output[i * 2 + 1] = f32(value * gain)
 
         update_phase(note, pDevice.sampleRate)
     }
